+++ docs/source/whats_new.rst
@@ -1,3 +1,22 @@
+What's New in cartopy 0.17
+==========================
+
+:Release: 0.17.0
+:Date:
+
+Features
+--------
+* The :class:`cartopy.feature.NaturalEarthFeature` class now allows a
+:class:`cartopy.feature.AdaptiveScaler` object to be passed as the `scale`
+argument. This will automatically choose the appropriate feature scale from
+the GeoAxes extent. This can also be used interactively while panning and
+zooming in a figure. :data:`cartopy.feature.NaturalEarthFeature.scale` is
+now read-only. (:pull:`1102`, :pull:`983`)
+
+--------
+
+
+
 What's New in cartopy 0.16
 ==========================
 
@@ -93,10 +112,8 @@ Features
 * If Fiona is installed on a user's system, this will now be the default
   shapefile reader, adding significant speed improvements. (:pull:`1000`)
 
-* The :class:`cartopy.feature.NaturalEarthFeature` class now allows the
-`scale` argument to be set to `'auto'`. This will automatically choose the
-appropriate feature scale from the GeoAxes extent. This can also be used
-interactively while panning and zooming in a figure.
+-----------
+
 
 
 What's New in cartopy 0.15